Update README.md
# Dictionary

Hello guys! this is my first ever kinda-useful-for-myself project I've ever made with Python!

What this file does is that when you enter a word inside that it uses a web scraping function that is connected to (https://www.dictionary.com/) and it finds the word's meaning and it automatically adds it to the internal dictionary that exists in the code, for saving it you will have to press " Right Shift " key on your keyboard.

Usage : Extract the py file into a folder, make another text file called " English ", run the python script and you are good to go.